requirements-dev.txt
-r requirements.txt
coverage
flake8
flake8-bugbear
flake8-docstrings
pep8-naming
pre-commit
pytest
requests-mock
